LPD-75968 portal-workflow-metrics-service: Extract BaseWorkflowMetricsReindexer template method
Lift the reindex(long, ExecutionMode) skeleton (guard, createMissingIndexes, SYNC/FULL prep, deleteStaleDocuments), the reindex(long) → FULL delegate, the isExecuteSyncReindex helper, the SyncReindexManager Snapshot, and the shared IndexNameBuilder/SearchCapabilities/SearchEngineAdapter references into an abstract BaseWorkflowMetricsReindexer.

Each of the seven reindexers now implements reindexEntities(long, ExecutionMode) for the entity-specific middle. A protected postReindex(long, ExecutionMode) hook (default no-op) leaves an extension point for downstream work.

Adds a public getIndexNameSuffix() accessor to WorkflowMetricsIndex so the base can compute the stale-doc index name from getKey() without a separate abstract.

https://liferay.atlassian.net/browse/LPD-75968
