+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-service/src/main/java/com/liferay/portal/workflow/metrics/internal/search/index/WorkflowMetricsIndex.java
@@ -138,6 +138,10 @@ public enum WorkflowMetricsIndex {
 		return true;
 	}
 
+	public String getIndexNameSuffix() {
+		return _indexNameSuffix;
+	}
+
 	public boolean removeIndex(
 			SearchCapabilities searchCapabilities,
 			SearchEngineAdapter searchEngineAdapter,
